Additional setup, helper functions, and policies
Adds helper functions in the setup migration for casting ts and
tstzranges to json objects. This allows clients of the Supabase API to
get intervals as this:

{"lower": "2022-07-10T13:42:16.566914+00:00", "upper": null}

instead of the usual Postgres representation of tstzranges.

The targets migration has been updated with a new policy allowing users
to delete their own targets.

The intervals migration has been updated with a delete policy as well.
Additionally a view is defined to get any active intervals which is
necessary as Postgrest's query functionality does not support range
operators on tstzrange columns. Convenience functions for starting and
stopping an interval are provided for ease of use.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>