Complete statistics views and functions + format
Adds complete statistics views and functions to get data for the
calendar heatmap and cumulative line chart statistics views. Both the
function and views are flexible enough to serve other use cases.

The views need to filter on the account variable as views in Postgres 14
are queried using the security privileges of the view creator rather
than the user running the query. This renders the row level policies of
the underlying tables ineffective. An option to activate RLS for views
is available in the upcoming Postgres 15 version which is as of yet
unreleased.

Use the Prettier SQL extension which use SQL Formatter under the hood:
https://github.com/sql-formatter-org/sql-formatter

This formatter can be added as a git hook for those that do not use
VSCode for editing.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>