Add json casts for timestamptz and intervals
Instead of trying to set the default interval and date styles to
iso_8601 we can add json casts for these value types instead. The users
of the Supabase backend then needs to type cast to json if they want
timestamptz and interval values as iso8601 strings.

For json aggregated columns the values will be automatically returned as
iso8601. This should fit the expectation that json aggregation returns
all contained values in json format.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>